frontend base integration (vue)

+++ src/Controller/API/Rest/CardController.php
@@ -1,18 +1,24 @@
 <?php
 
-namespace App\Controller\Rest;
+declare(strict_types=1);
 
-use App\Controller\BaseController;
+namespace App\Controller\API\Rest;
+
+use App\Controller\API\BaseController;
 use App\Entity\Card;
+use App\Entity\Deck;
 use App\Model\CardDTO;
+use FOS\RestBundle\Controller\Annotations as Rest;
 use FOS\RestBundle\Routing\ClassResourceInterface;
 use FOS\RestBundle\View\View;
 use JMS\Serializer\SerializerInterface;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
+use Swagger\Annotations as SWG;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use FOS\RestBundle\Controller\Annotations as Rest;
-use Swagger\Annotations as SWG;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
+use function array_merge;
+use function assert;
 
 /**
  * @Rest\Prefix(value="api/v1")
@@ -27,18 +33,19 @@ class CardController extends BaseController implements ClassResourceInterface
         $this->serializationGroup = ['card_list'];
     }
 
-
     /**
+     * @return mixed
+     *
      * @SWG\Response(
      *      response="200",
      *      description="Success",
      * )
      * @SWG\Tag(name="CardController")
-     * @param Card $card
-     * @return mixed
      */
     public function getAction(Card $card)
     {
+        $this->denyAccessUnlessGranted('view', $card);
+
         return $this->viewSerialized(
             $card,
             array_merge($this->serializationGroup, ['card_details'])
@@ -46,19 +53,21 @@ class CardController extends BaseController implements ClassResourceInterface
     }
 
     /**
+     * @return mixed
+     *
+     * @Rest\QueryParam()
      * @SWG\Response(
      *      response="200",
      *      description="Success",
      * )
      * @SWG\Tag(name="CardController")
-     * @return mixed
      */
     public function cgetAction()
     {
         $cards = $this
             ->getDoctrine()
             ->getRepository(Card::class)
-            ->findAll();
+            ->findBy(['user' => $this->getUser()]);
 
         return $this->viewSerialized($cards, $this->serializationGroup);
     }
@@ -68,36 +77,42 @@ class CardController extends BaseController implements ClassResourceInterface
      *      response="200",
      *      description="Success",
      * )
+     * @Rest\Post(path="deck/{deck}/cards/")
+     * @ParamConverter(name="deck", class="App\Entity\Deck")
      * @SWG\Tag(name="CardController")
-     * @param Request $request
-     * @return View
      */
-    public function postAction(Request $request)
+    public function postAction(Request $request, Deck $deck) : View
     {
-        /** @var CardDTO $cardDTO */
+        $this->denyAccessUnlessGranted('edit', $deck);
+
         $cardDTO = $this->validateRequestData($request, CardDTO::class);
+        assert($cardDTO instanceof CardDTO);
         $card = $cardDTO->fromDTO();
-        $this->fastSave($card);
-        return $this->view(["status" => "ok"], Response::HTTP_CREATED);
+        $card->setDeck($deck);
+        $this->persist($card);
+
+        return $this->view(['status' => 'ok'], Response::HTTP_CREATED);
     }
 
     /**
+     * @param Deck $deck
+     *
      * @SWG\Response(
      *      response="200",
      *      description="Success",
      * )
      * @SWG\Tag(name="CardController")
-     * @param Request $request
-     * @param Card $card
-     * @return View
+     * @ParamConverter(name="deck", class="App\Entity\Deck")
      */
-    public function putAction(Request $request, Card $card)
+    public function putAction(Request $request, Card $card) : View
     {
-        /** @var CardDTO $cardDTO */
+        $this->denyAccessUnlessGranted('edit', $card->getDeck());
         $cardDTO = $this->validateRequestData($request, CardDTO::class);
+        assert($cardDTO instanceof CardDTO);
         $card = $cardDTO->fromDTO($card);
         $this->getDoctrine()->getManager()->flush();
-        return $this->view(["status" => "success"]);
+
+        return $this->view(['status' => 'success']);
     }
 
     /**
@@ -106,17 +121,16 @@ class CardController extends BaseController implements ClassResourceInterface
      *      description="Success",
      * )
      * @SWG\Tag(name="CardController")
-     * @param Request $request
-     * @param Card $card
-     * @return View
      */
-    public function patchAction(Request $request, Card $card)
+    public function patchAction(Request $request, Card $card) : View
     {
-        /** @var CardDTO $cardDTO */
+        $this->denyAccessUnlessGranted('edit', $card->getDeck());
         $cardDTO = $this->validateRequestData($request, CardDTO::class);
+        assert($cardDTO instanceof CardDTO);
         $card = $cardDTO->fromDTO($card);
         $this->getDoctrine()->getManager()->flush();
-        return $this->view(["status" => "success"]);
+
+        return $this->view(['status' => 'success']);
     }
 
     /**
@@ -125,14 +139,14 @@ class CardController extends BaseController implements ClassResourceInterface
      *      description="Success",
      * )
      * @SWG\Tag(name="CardController")
-     * @param Card $card
-     * @return View
      */
-    public function deleteAction(Card $card)
+    public function deleteAction(Card $card) : View
     {
+        $this->denyAccessUnlessGranted('edit', $card->getDeck());
         $em = $this->getDoctrine()->getManager();
         $em->remove($card);
         $em->flush();
+
         return $this->view(['success' => true]);
     }
 }
